+++ src/main.lua
@@ -1,5 +1,7 @@
 local function Main()
     -- require 'test.framework.behavior'
+    print(string.gmatch("烈焰-火焰強化", '.+'))
+    print(string.match("火焰強化", '-.+'))
 end
 
 Main()